TGA->JPEG compression: added support for transparency.
The alpha channel is saved as imagename_alpha.jpg, but only if it exists and isn't completely white.
All versions of DPRM support this.

In addition, the compressdirs file is now ignored by default, and all TGAs found in the package are compressed, with a few specific exceptions where this causes problems.